Move udp get_fd code int netex
Also print ip for tcp.
