+++ tcp_file_transfer_client.c
@@ -121,9 +121,6 @@ int
 main(int argc, char *argv[])
 {
   int opt;
-  conn_info conn_inf;
-  conn_inf.host = default_host;
-  conn_inf.port = default_port;
   char *file = NULL;
 
   while ((opt = getopt(argc, argv, "f:a:p:h")) != -1)
@@ -152,7 +149,7 @@ main(int argc, char *argv[])
     exit(EXIT_FAILURE);
   }
 
-  int res = tcp_client_conn(&conn_inf);
+  int res = get_tcp_client_sockfd();
   if (res < 0)
     return res;
